improv: remove root logger logic, lib
